+++ app/src/main/java/com/igorganapolsky/vibratingwatchapp/domain/manager/vibration/BeepManager.java
@@ -1,4 +1,4 @@
-package com.igorganapolsky.vibratingwatchapp.manager.vibration;
+package com.igorganapolsky.vibratingwatchapp.domain.manager.vibration;
 
 import com.igorganapolsky.vibratingwatchapp.domain.model.BuzzSetup;
