+++ app/src/main/java/com/igorganapolsky/vibratingwatchapp/domain/model/BuzzSetup.java
@@ -2,11 +2,11 @@ package com.igorganapolsky.vibratingwatchapp.domain.model;
 
 public class BuzzSetup {
 
-   public enum Type {SHORT, LONG}
+    public enum Type {SHORT, LONG}
 
-    private Type buzzType;
-    private int buzzCount;
-    private int buzzTime;
+    private final Type buzzType;
+    private final int buzzCount;
+    private final int buzzTime;
 
     public BuzzSetup(Type buzzType, int buzzCount, int buzzTime) {
         this.buzzType = buzzType;
