+++ app/src/main/java/com/igorganapolsky/vibratingwatchapp/presentation/main/TimerListViewModel.java
@@ -4,8 +4,8 @@ import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.Observer;
 import android.arch.lifecycle.ViewModel;
 import com.igorganapolsky.vibratingwatchapp.domain.Repository;
+import com.igorganapolsky.vibratingwatchapp.domain.manager.timer.CountdownManager;
 import com.igorganapolsky.vibratingwatchapp.domain.model.TimerModel;
-import com.igorganapolsky.vibratingwatchapp.manager.timer.CountdownManager;
 
 import java.util.List;
